popovers: Move navbar popovers to a subdirectory.
