+++ web/templates/popovers/navbar/navbar_gear_menu_popover.hbs
