+++ web/templates/popovers/navbar/navbar_personal_menu_popover.hbs
